// frontend/ticketing_app_ui/src/app/services/models/category-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { SupportServiceResponse } from '../models/support-service-response';
export interface CategoryResponse {
  description?: string;
  id?: number;
  name?: string;
  services?: Array<SupportServiceResponse>;
  ticketCount?: number;
}