// frontend/ticketing_app_ui/src/app/services/models/ticket-response-dto.ts
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface TicketResponseDto {
  assignedToId?: string;
  assignedToName?: string;
  categoryId?: number;
  categoryName?: string;
  createdOn?: string;
  description?: string;
  id?: number;
  priority?: 'LOW' | 'MEDIUM' | 'HIGH' | 'CRITICAL';
  solvedOn?: string;
  status?: 'OPEN' | 'ANSWERED' | 'SOLVED' | 'DRAFT';
  supportServiceId?: number;
  supportServiceName?: string;
  title?: string;
  updatedOn?: string;
  userEmail?: string;
  userFirstName?: string;
  userFiscalCode?: string;
  userId?: string;
  userLastName?: string;
  userPhoneNumber?: string;
}